refactor([issue-749]): extract state, stream-parse, and process-stats layers from cos-runner
server/cos-runner/index.js had grown to 1113 LOC mixing four concerns in one
file. Split the standalone portos-cos runner into focused sibling modules,
keeping the directory self-contained (no imports out to the main server's
dependency graph) so the isolated PM2 process stays standalone:

- streamJsonParser.js — safeParse, summarizeToolInput, createStreamJsonParser
- runnerState.js — DEFAULT_STATE, atomicWrite, loadState, saveState
- processStats.js — getProcessStats, checkProcessRunning (PID-validated)

index.js drops to 835 LOC and now only owns the express/socket.io IPC surface
and agent/run lifecycle wiring. Added unit tests for the extracted pure
helpers, which previously had no coverage.
